requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==2.2.4
django-appconf==1.0.3
django-bootstrap4==0.0.8
django-cors-headers==3.1.0
django-heroku==0.3.1
django-model-utils==3.2.0
django-rest-swagger==2.2.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==6.1.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycparser==2.19
PyJWT==1.7.1
python-decouple==3.1
python-http-client==3.1.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
whitenoise==4.1.3
